Add unified backend sample, update pr, d3d12
